........
  r581341 | dkulp | 2007-10-02 14:31:33 -0400 (Tue, 02 Oct 2007) | 7 lines
  
  [CXF-1060]
  * Workaround bug in Spring where Maps injected into contructors are always copied into LinkedHashMaps thus loosing concurrency contraints as well as causing all "delayed" beans to be creates/resolved immediately.  (This includes bindings, conduits, destinations, and a bunch of ws-policy assertion things).
  * Create ClientOnlyHTTPTransportFactory for use with clients that don't need ws-a style callback things.  (this is part 1.   There are still areas in the code that check for transport Id's from the destinations instead of conduits.   In this case, there will not be a destination registered.)
  * For our tests, don't close down the port.  It's causing all kinds of hangs on my Linux box.  (probably bug in JDK or jetty)
  * Update to latest gpg plugin
........


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/branches/2.0.x-fixes@582920 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/transports/http-jetty/src/main/java/org/apache/cxf/transport/http_jetty/JettyHTTPServerEngine.java
@@ -165,11 +165,28 @@ public class JettyHTTPServerEngine
      * remove it from the factory's cache. 
      */
     public void shutdown() {
-        if (servantCount == 0) {
-            factory.destroyForPort(port);
-        } else {
-            LOG.log(Level.WARNING, "FAILED_TO_SHOWDOWN_ENGINE_MSG", port);
+        if (shouldDestroyPort()) {
+            if (servantCount == 0) {
+                factory.destroyForPort(port);
+            } else {
+                LOG.log(Level.WARNING, "FAILED_TO_SHOWDOWN_ENGINE_MSG", port);
+            }
         }
+    }
+    
+    private boolean shouldDestroyPort() {
+        //if we shutdown the port, on SOME OS's/JVM's, if a client
+        //in the same jvm had been talking to it at some point and keep alives
+        //are on, then the port is held open for about 60 seconds
+        //afterword and if we restart, connections will then 
+        //get sent into the old stuff where there are 
+        //no longer any servant registered.   They pretty much just hang.
+        
+        //this is most often seen in our unit/system tests that 
+        //test things in the same VM.
+        
+        String s = System.getProperty("org.apache.cxf.transports.http_jetty.DontClosePort");
+        return !Boolean.valueOf(s);
     }
     
     /**
@@ -475,6 +492,8 @@ public class JettyHTTPServerEngine
         }
     }
     
+
+    
     /**
      * This method is called by the ServerEngine Factory to destroy the 
      * listener.
@@ -482,10 +501,11 @@ public class JettyHTTPServerEngine
      */
     protected void stop() throws Exception {
         if (server != null) {
+            
             connector.close();
             server.stop();
             server.destroy();
-            server   = null;
+            server = null;
         }
     }
